Fix bug: UI
By storing the letter-items ID and not the entire letter-item object itself in
state, the `isHighlighted` prop, responsible for the "highlighted" className
logic, returns the correct value for the intended UI list-item highlighting
interaction. The previous unintended interaction was due to only one of two
duplicate objects being updated when the user clicked the star/unstar button.

Deduping state prevents the expression passed by the `isHighlighted` prop from
becoming false when the `handleStar` handler toggles the items `isStarred`
property.

In short, we created a duplicate state and forgot to update both items.
And instead of updating both items, we chose to remove the duplicate and
alternatively base the highlighting logic on the ID.